added overflow icon to each listview card. added a delete menu to it. doesn't do anything yet. also added a confirmation dialog for the user to confirm deleting all the records. added a few checks there as well to make sure it only works when the list is populated.

+++ SimpleMPGCalculator/src/com/jackson/simplempgcalculator/PastResults.java
@@ -1,14 +1,19 @@
 package com.jackson.simplempgcalculator;
 
+import android.app.AlertDialog;
 import android.app.Fragment;
+import android.content.DialogInterface;
 import android.database.Cursor;
 import android.os.Bundle;
 import android.view.LayoutInflater;
+import android.view.MenuInflater;
 import android.view.View;
 import android.view.View.OnClickListener;
 import android.view.ViewGroup;
 import android.widget.Button;
+import android.widget.ImageButton;
 import android.widget.ListView;
+import android.widget.PopupMenu;
 import android.widget.SimpleCursorAdapter;
 import android.widget.Toast;
 
@@ -16,6 +21,7 @@ public class PastResults extends Fragment implements OnClickListener {
 	
 	/* VARIABLES */
 	private Button delete;
+	public Boolean listIsEmpty = true;
 
 	@Override
 	public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
@@ -43,12 +49,19 @@ public class PastResults extends Fragment implements OnClickListener {
 		String selectAll = "SELECT * FROM trips";
 		Cursor cursor = adapter.select(selectAll);
 		
+		if(cursor.moveToFirst()) {
+			listIsEmpty = false;
+		} else {
+
+			listIsEmpty = true;
+		}
+		
 		String from[] = new String[]
 				{DbAdapter.GALLONS, DbAdapter.MILES, DbAdapter.MPG,
-					DbAdapter.PRICE, DbAdapter.TOTAL_COST, DbAdapter.DATE};
+					DbAdapter.PRICE, DbAdapter.TOTAL_COST, DbAdapter.DATE, DbAdapter.ID};
 		int[] to = new int[]
 				{R.id.gallons_item, R.id.miles_item, R.id.mpg_item,
-					R.id.price_item, R.id.cost_item, R.id.date};
+					R.id.price_item, R.id.cost_item, R.id.date, R.id.rowId};
 		
 		
 		SimpleCursorAdapter myCursorAdapter = new SimpleCursorAdapter(getActivity(), R.layout.result_item, cursor, from, to);
@@ -64,26 +77,56 @@ public class PastResults extends Fragment implements OnClickListener {
 	public void onClick(View v) {
 		switch(v.getId()) {
 		case R.id.delete_all:
-			deleteAllRecords();
+			//check to see if there are any records in the listview before proceeding
+			if(listIsEmpty) {
+				Toast.makeText(getActivity(), "There are no records to delete", Toast.LENGTH_SHORT).show();
+			} else {
+				deleteAllRecords();
+			}
 			break;
 		}	
 	}
 
-	private void deleteAllRecords() {
-		DbAdapter adapter = new DbAdapter(getActivity());
-		
-		adapter.open();
-		
-		int rows = adapter.delete();
-		
-		if(rows <= 0){
-			Toast.makeText(getActivity(), "No rows were deleted", Toast.LENGTH_SHORT).show();
-		} else {
-			Toast.makeText(getActivity(), "All trips successfully deleted", Toast.LENGTH_SHORT).show();
-		}
-		
-		populateResultsList();
+	private void deleteAllRecords() {		
+		new AlertDialog.Builder(getActivity())
+	        .setIcon(android.R.drawable.ic_dialog_alert)
+	        .setTitle(R.string.deleteTitle)
+	        .setMessage(R.string.deleteMessage)
+	        .setPositiveButton(R.string.deleteYes, new DialogInterface.OnClickListener() {
+		        	@Override
+				    public void onClick(DialogInterface dialog, int which) {
+		        		DbAdapter adapter = new DbAdapter(getActivity());
+		        		
+		        		adapter.open();
+		        		
+		        		int rows = adapter.delete();
+		        		
+		        		if(rows <= 0){
+		        			Toast.makeText(getActivity(), "No rows were deleted", Toast.LENGTH_SHORT).show();
+		        		} else {
+		        			Toast.makeText(getActivity(), "All trips successfully deleted", Toast.LENGTH_SHORT).show();
+		        		}
+		        		
+		        		populateResultsList();
+			        }
+			})
+		    .setNegativeButton(R.string.deleteNo, null)
+		    .show();
 	}
 
-	
 }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
